% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{backwd}
\alias{backwd}
\title{Backward Algorithm}
\usage{
backwd(obs, transition, emission)
}
\arguments{
\item{obs}{Vector of visible observations}

\item{transition}{Transition Matrix}

\item{emission}{Emission matrix}
}
\value{
beta matrix of probabilities of model being in state s at time t
}
\description{
Calculates the probability of the sequence ending given state i at time t
}
