% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{BaumWelch}
\alias{BaumWelch}
\title{Baum-Welch algorithm}
\usage{
BaumWelch(visible, t.mat, e.mat, initial_distribution, n.iter = 100)
}
\arguments{
\item{visible}{Vector of visible observations}

\item{t.mat}{Transition Matrix}

\item{e.mat}{Emission matrix}
}
\value{
list containing a: transition matrix, b: emission matrix, initial_distribution vector
}
\description{
Updates transition and emission matrices using forward-backward algorithm
}
