% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{create_sequence}
\alias{create_sequence}
\title{Create sequence from HMM}
\usage{
create_sequence(transition.mat, emission.mat, n, start.prob)
}
\arguments{
\item{transition.mat}{Transition Matrix specifying the probability of changing or remaining in the same state}

\item{emission.mat}{Emission matrix specifying probability of emitting certain symbol from a state}

\item{n}{The length of the sequence to be generated}

\item{start.prob}{Vector of initial probabilities, probaility of starting from a state}
}
\value{
list containing generated sequence and the vector of states under which each symbol was generated
}
\description{
Create a sequence from a simulated HMM given transition and emission matrices
}
