% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{forward}
\alias{forward}
\title{Forward algorithm}
\usage{
forward(obs, transition, E)
}
\arguments{
\item{obs}{Vector of visible observations}

\item{transition}{Transition matrix, columns must sum 1}

\item{E}{Emission matrix, rows must sum 1}
}
\value{
alpha matrix of probabilities of being in a certain state for each step
}
\description{
Calculates the probability of the next emission, using the computed probability of the current emission
}
